// Capstone/C9-Pavithra-HMS-Frontend-Project-S1n/Frontend/Hotel Management System/hms/hms/src/app/models/guests.model.ts
// export class Guests {
//     id: string = "";
//     name: string = ""; 
//     email: string = "";
//     phone: string = "";
//     nationality: string = "";
//   hotel: any;
//   }
export class Guests {
  id: string="";
  name: string="";
  email: string="";
  phone: string="";
  nationality: string="";
 
}